feat: add temp dir and dont delete  gitkeep

+++ src/bin/www.mjs
@@ -30,7 +30,9 @@ const cleanupImage = async () => {
 
     res.map(async (fileName) => {
         const filePath = `${path.join(__dirname, '..', 'temp', 'images')}/${fileName}`;
-        console.log(filePath);
+
+        if (fileName === '.gitkeep') return;
+        
         await unlink(filePath);
     });
 };
